requirements.txt
pydantic==2.5.0
numpy==1.24.3
pandas==2.0.3
jsonschema==4.17.3
python-json-logger==2.0.7
asyncio==3.4.3
aiofiles==23.2.1
fastapi==0.104.1
uvicorn==0.24.0
psutil==5.9.5
sqlalchemy==2.0.23
redis==5.0.1
# ML and Scientific Computing
scikit-learn==1.3.0
torch==2.1.0
transformers==4.35.0
scipy==1.11.4
matplotlib==3.7.2
seaborn==0.12.2
# Database
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0
# Web and API
jinja2==3.1.2
python-multipart==0.0.6
# Monitoring and Metrics
prometheus-client==0.19.0
# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
# Security and Authentication
PyJWT==2.8.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
# Advanced Analytics
plotly==5.17.0
dash==2.14.2
jupyter==1.0.0
ipython==8.17.2
# Production dependencies
gunicorn==21.2.0
celery==5.3.4
# Additional utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0
# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
pre-commit==3.5.0
structlog==23.2.0
colorlog==6.8.0
# Data Processing
openpyxl==3.1.2
xlsxwriter==3.1.9
# Additional production dependencies
aiohttp==3.9.1
websockets==12.0
